--- apps/src/CDOIDE/cdoIDEContext/index.ts
@@ -1,3 +1,0 @@
-export * from './cdoIDEContext';
-export * from './cdoIDEContextProjectReducer';
-export * from './utils';
