--- apps/src/CDOIDE/hooks/useEmptyEditor.tsx
@@ -1,7 +1,0 @@
-import {useCDOIDEContext} from '@cdoide/cdoIDEContext';
-import {getEmptyEditor} from '@cdoide/utils';
-
-export const useEmptyEditor = () => {
-  const {config} = useCDOIDEContext();
-  return getEmptyEditor(config);
-};
