--- apps/src/CDOIDE/index.ts
@@ -1,5 +1,0 @@
-export * from './CDOIDE';
-export * from './types';
-export * from './cdoIDEContext';
-export * from './cdoIDEContext/types';
-export * from './utils';
