+++ apps/src/codebridge/Editor/styles/editor.css
