+++ apps/src/codebridge/codebridgeContext/codebridgeContextProjectReducer.ts
@@ -1,5 +1,10 @@
-import {ProjectType, ReducerAction, FileId, FolderId} from '@cdoide/types';
-import {sortFilesByName} from '@cdoide/utils';
+import {
+  ProjectType,
+  ReducerAction,
+  FileId,
+  FolderId,
+} from '@cdo/apps/codebridge/types';
+import {sortFilesByName} from '@cdo/apps/codebridge/utils';
 
 import {PROJECT_REDUCER_ACTIONS} from './constants';
 import {findFiles, findSubFolders} from './utils';
