+++ apps/src/codebridge/codebridgeContext/types.ts
@@ -1,4 +1,4 @@
-import {ProjectType, FileId, FolderId} from '@cdoide/types';
+import {ProjectType, FileId, FolderId} from '@cdo/apps/codebridge/types';
 export type ReplaceProjectFunction = (project: ProjectType) => void;
 
 export type SaveFileFunction = (fileId: FileId, contents: string) => void;
