+++ apps/src/codebridge/hooks/useEmptyEditor.tsx
@@ -1,0 +1,7 @@
+import {useCDOIDEContext} from '@cdo/apps/codebridge/codebridgeContext';
+import {getEmptyEditor} from '@cdo/apps/codebridge/utils';
+
+export const useEmptyEditor = () => {
+  const {config} = useCDOIDEContext();
+  return getEmptyEditor(config);
+};
