+++ apps/src/codebridge/utils/getEmptyProject.ts
