+++ apps/src/pythonlab/PythonlabView.tsx
@@ -1,11 +1,11 @@
 // Pythonlab view
 import React, {useEffect, useMemo, useState} from 'react';
 import moduleStyles from './pythonlab-view.module.scss';
-import {ConfigType} from '@cdo/apps/CDOIDE/types';
-import {Editor} from '@cdo/apps/CDOIDE/Editor';
+import {ConfigType} from '@cdo/apps/codebridge/types';
+import {Editor} from '@cdo/apps/codebridge/Editor';
 import {LanguageSupport} from '@codemirror/language';
 import {python} from '@codemirror/lang-python';
-import {CDOIDE} from '@cdo/apps/CDOIDE';
+import {CDOIDE} from '@cdo/apps/codebridge';
 import {MultiFileSource} from '@cdo/apps/lab2/types';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 import {setAndSaveSource, setSource} from './pythonlabRedux';
@@ -13,7 +13,7 @@ import PythonConsole from './PythonConsole';
 import {MAIN_PYTHON_FILE, START_SOURCES} from '@cdo/apps/lab2/constants';
 import {getAppOptionsEditBlocks} from '@cdo/apps/lab2/projects/utils';
 import header from '@cdo/apps/code-studio/header';
-import {useInitialSources} from '@cdoide/hooks';
+import {useInitialSources} from '@cdo/apps/codebridge/hooks';
 
 const pythonlabLangMapping: {[key: string]: LanguageSupport} = {
   py: python(),
