+++ apps/src/pythonlab/pythonHelpers/pythonScriptUtils.ts
@@ -1,8 +1,11 @@
-import {getNextFileId, getNextFolderId} from '@cdoide/cdoIDEContext/utils';
-import {DEFAULT_FOLDER_ID} from '@cdoide/constants';
 import _ from 'lodash';
 import {PyodideInterface} from 'pyodide';
 
+import {
+  getNextFileId,
+  getNextFolderId,
+} from '@cdo/apps/codebridge/codebridgeContext/utils';
+import {DEFAULT_FOLDER_ID} from '@cdo/apps/codebridge/constants';
 import {MultiFileSource} from '@cdo/apps/lab2/types';
 
 import {PyodideMessage, PyodidePathContent} from '../types';
